added clearForm function and moved functions to index.js
