+++ src/index.js
@@ -1,6 +1,85 @@
 import './style.css';
 import todo from './todo';
-import { project, save, setActiveProjectId, getActiveProject, render } from './project';
+import { project, projects } from './project';
+
+let activeProjectId = JSON.parse(localStorage.getItem('active_project_id_key'));
+
+const renderTodos = () => {
+const activeProject = getActiveProject();
+const todosDiv = document.querySelector('.todos');
+clearContainer(todosDiv);
+activeProject.todos.forEach((todo, idx) => {
+  let template = document.createElement("a");
+  template.setAttribute('href','#');
+  template.setAttribute('id', `${ idx }`);
+  template.classList.add("list-group-item", "list-group-item-action");
+  template.innerHTML = `<h4>${ todo.title }</h4>`;
+  let details = document.createElement("p");
+  details.classList.add("hidden");
+  details.innerText = "Description here \n Description here";
+  template.appendChild(details)
+  let editTodo = document.createElement("div");
+  editTodo.classList.add("edit-todo", "text-right");
+  editTodo.innerHTML = `
+  <span><button class="btn btn-success">&#x2714;</button></span>
+  <span><button class="btn btn-danger">&#x2715;</button></span>
+  <span><button class="btn btn-primary">&#x270E;</button></span>`;
+  template.appendChild(editTodo);
+  if (todo.completed) {
+    template.classList.add('completed');
+  }
+  todosDiv.appendChild(template);
+});
+};
+
+const renderProjects = () => {
+  const projectsDiv = document.querySelector('.project');
+  clearContainer(projectsDiv);
+  projects.forEach((project) => {
+    let template = document.createElement("a");
+    template.classList.add("list-group-item", "list-group-item-action");
+    template.setAttribute("data-id", `${project.id}`);
+    template.setAttribute('href','#');
+    template.innerHTML = `${ project.name } <span class="badge badge-primary badge-pill">${ project.todos.length }</span>`;
+    if (project.id === activeProjectId) {
+      template.classList.add("active");
+    }
+    projectsDiv.appendChild(template);
+  });
+};
+
+const render = () => {
+  renderProjects();
+  renderTodos();
+};
+
+const save = () => {
+  localStorage.setItem('projects_key', JSON.stringify(projects));
+  localStorage.setItem('active_project_id_key', JSON.stringify(activeProjectId));
+  render();
+};
+
+const setActiveProjectId = (value) => {
+  activeProjectId = value;
+};
+
+const getActiveProject = () => {
+  return projects.find(project => project.id === activeProjectId);
+};
+
+const clearContainer = (container) => {
+  while(container.firstChild) {
+    container.firstChild.remove();
+  }
+};
+
+const clearForm = (form) => {
+  document.querySelector('#title').value = '';
+  document.querySelector('#desc').value = '';
+  document.querySelector('#date').value = '';
+  document.querySelector('#projectName').value = '';
+  form.classList.toggle('hidden');
+};
 
 const pageLoad = () => {
   const newProjectBtn = document.querySelector('.new-project');
@@ -9,10 +88,10 @@ const pageLoad = () => {
   const newTodoForm = document.querySelector('.new-todo-form');
   const projectsContainer = document.querySelector('.project');
   const todosContainer = document.querySelector('.list');
-  const editTodo = document.querySelector('.edit-todo')
+  const editTodo = document.querySelector('.edit-todo');
 
   newProjectBtn.addEventListener('click', () => {
-    newProjectForm.classList.toggle('hidden');
+    clearForm(newProjectForm);
   });
 
   newProjectForm.addEventListener('submit', (event) => {
@@ -21,12 +100,13 @@ const pageLoad = () => {
     if ( name !== "") {
       const newProject = project(name);
       newProject.addToProjects();
+      clearForm(newProjectForm);
       save();
     } 
   });
 
   newTodoBtn.addEventListener('click', () => {
-    newTodoForm.classList.toggle('hidden');
+    clearForm(newTodoForm);
   });
 
   newTodoForm.addEventListener('submit', (event) => {
@@ -40,6 +120,7 @@ const pageLoad = () => {
       const newTodo = todo(title, desc, date, priority);
       let activeProject = getActiveProject();
       activeProject.todos.push(newTodo);
+      clearForm(newTodoForm);
       save();
     }
   });
